8_RDBMS/19_PYthon_MySQL/requirements.txt
astroid==3.0.2
colorama==0.4.6
dill==0.3.7
isort==5.13.2
mccabe==0.7.0
mysql-connector-python==9.0.0
platformdirs==4.1.0
pylint==3.0.3
setuptools==72.1.0
tomlkit==0.12.3
wheel==0.43.0
